requirements.txt
asgiref==3.7.2
Django==5.0.2
pillow==10.2.0
sqlparse==0.4.4
tzdata==2024.1
requests
gunicorn==21.2.0
packaging==23.2
whitenoise
